+++ src/main/java/com/hhconcert/server/business/domain/queues/service/TokenActiveScheduler.java
@@ -10,14 +10,11 @@ import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.time.LocalDateTime;
-import java.util.List;
-
 @Slf4j
 @EnableScheduling
 @Component
 @RequiredArgsConstructor
-public class TokenScheduler {
+public class TokenActiveScheduler {
 
     private final static int MAX_ACTIVE_TOKEN_COUNT = 10;
     private final TokenRepository tokenRepository;
@@ -38,13 +35,4 @@ public class TokenScheduler {
         }
     }
 
-    @Scheduled(cron = "0 * * * * *") // 1분마다 실행
-    @Transactional
-    public void dropTokens() {
-        List<Token> expiredTokens = tokenRepository.getExpiredTokens(LocalDateTime.now());
-        for (Token token : expiredTokens) {
-            tokenRepository.dropToken(token);
-        }
-        log.info("remove expired tokens : {}", expiredTokens.size());
-    }
 }
