+++ src/test/java/com/hhconcert/server/business/domain/queues/service/TokenExpireSchedulerTest.java
@@ -4,8 +4,6 @@ import com.hhconcert.server.business.domain.queues.entity.Token;
 import com.hhconcert.server.business.domain.queues.entity.TokenGenerator;
 import com.hhconcert.server.business.domain.queues.entity.TokenStatus;
 import com.hhconcert.server.business.domain.queues.persistance.TokenRepository;
-import com.hhconcert.server.infrastructure.queues.TokenJpaRepository;
-import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -15,41 +13,18 @@ import org.springframework.test.context.ActiveProfiles;
 import java.time.LocalDateTime;
 import java.util.NoSuchElementException;
 
-import static org.assertj.core.api.Assertions.*;
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 @ActiveProfiles("test")
 @SpringBootTest
-class TokenSchedulerTest {
+class TokenExpireSchedulerTest {
 
     @Autowired
-    private TokenScheduler tokenScheduler;
+    private TokenExpireScheduler tokenExpireScheduler;
 
     @Autowired
     private TokenRepository tokenRepository;
-
-    @Autowired
-    private TokenJpaRepository tokenJpaRepository;
-
-    @AfterEach
-    void tearDown() {
-        tokenJpaRepository.deleteAllInBatch();
-    }
-
-    @DisplayName("대기열 통과 시, WAIT 상태 토큰 1개를 ACTIVE 상태로 변경한다.")
-    @Test
-    void whenPassedQueue_thenUpdateStatusIsActivate() {
-        Token token1 = tokenRepository.createToken(Token.createForWait("test1", 1));
-        Token token2 = tokenRepository.createToken(Token.createForWait("test2", 2));
-
-        tokenScheduler.activateTokens();
-
-        Token findToken1 = tokenRepository.findToken(token1.getTokenId());
-        assertThat(findToken1.getStatus()).isEqualTo(TokenStatus.ACTIVE);
-
-        Token findToken2 = tokenRepository.findToken(token2.getTokenId());
-        assertThat(findToken2).extracting("status", "priority")
-                        .containsExactly(TokenStatus.WAIT, 1);
-    }
 
     @DisplayName("만료된 ACTIVE 상태의 토큰을 삭제한다.")
     @Test
@@ -60,7 +35,7 @@ class TokenSchedulerTest {
         Token token2 = tokenRepository.createToken(new Token(TokenGenerator.generateToken("test2"), "test2",
                 TokenStatus.ACTIVE, 0, now.minusMinutes(2), now.plusMinutes(3)));
 
-        tokenScheduler.dropTokens();
+        tokenExpireScheduler.dropTokens();
 
         assertThatThrownBy(() -> tokenRepository.findToken(token1.getTokenId()))
                 .isInstanceOf(NoSuchElementException.class)
@@ -69,4 +44,5 @@ class TokenSchedulerTest {
         Token dropToken2 = tokenRepository.findToken(token2.getTokenId());
         assertThat(dropToken2).isNotNull();
     }
+
 }
